+++ Control Bibliotecario/Control Bibliotecario.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.34114.132
+# Visual Studio Version 17
+VisualStudioVersion = 17.7.34221.43
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Control Bibliotecario", "Control Bibliotecario\Control Bibliotecario.csproj", "{C2387EBF-E543-4521-9E02-B3C09B075DB8}"
 EndProject
