+++ Control Bibliotecario/Control Bibliotecario/Control Bibliotecario.csproj
@@ -46,33 +46,33 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Administrador.cs" />
-    <Compile Include="IngresarForm.cs">
+    <Compile Include="Modelos\Administrador.cs" />
+    <Compile Include="UI\IngresarForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="IngresarForm.Designer.cs">
+    <Compile Include="UI\IngresarForm.Designer.cs">
       <DependentUpon>IngresarForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="PrincipalForm.cs">
+    <Compile Include="UI\PrincipalForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include=".Designer.cs">
+    <Compile Include="UI\.Designer.cs">
       <DependentUpon>PrincipalForm.cs</DependentUpon>
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RegistroForm.cs">
+    <Compile Include="UI\RegistroForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="RegistroForm.Designer.cs">
+    <Compile Include="UI\RegistroForm.Designer.cs">
       <DependentUpon>RegistroForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="Usuario.cs" />
-    <EmbeddedResource Include="IngresarForm.resx">
+    <Compile Include="Modelos\Usuario.cs" />
+    <EmbeddedResource Include="UI\IngresarForm.resx">
       <DependentUpon>IngresarForm.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="PrincipalForm.resx">
+    <EmbeddedResource Include="UI\PrincipalForm.resx">
       <DependentUpon>PrincipalForm.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
@@ -84,7 +84,7 @@
       <AutoGen>True</AutoGen>
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
-    <EmbeddedResource Include="RegistroForm.resx">
+    <EmbeddedResource Include="UI\RegistroForm.resx">
       <DependentUpon>RegistroForm.cs</DependentUpon>
     </EmbeddedResource>
     <None Include="Properties\Settings.settings">
