+++ Control Bibliotecario/Control Bibliotecario/Modelos/Administrador.cs
@@ -1,0 +1,6 @@
+﻿namespace Control_Bibliotecario
+{
+    class Administrador
+    {
+    }
+}
