+++ Control Bibliotecario/Control Bibliotecario/Modelos/Usuario.cs
@@ -1,0 +1,6 @@
+﻿namespace Control_Bibliotecario
+{
+    class Usuario
+    {
+    }
+}
