+++ Control Bibliotecario/Control Bibliotecario/Program.cs
@@ -1,7 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using System.Windows.Forms;
 
 namespace Control_Bibliotecario
@@ -16,7 +13,7 @@ namespace Control_Bibliotecario
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new Form1());
+            Application.Run(new PrincipalForm());
         }
     }
 }
