+++ Control Bibliotecario/Control Bibliotecario/UI/IngresarForm.cs
@@ -1,0 +1,12 @@
+﻿using System.Windows.Forms;
+
+namespace Control_Bibliotecario
+{
+    public partial class IngresarForm : Form
+    {
+        public IngresarForm()
+        {
+            InitializeComponent();
+        }
+    }
+}
