+++ Control Bibliotecario/Control Bibliotecario/UI/PrincipalForm.cs
@@ -1,18 +1,11 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
 
 namespace Control_Bibliotecario
 {
-    public partial class Form1 : Form
+    public partial class PrincipalForm : Form
     {
-        public Form1()
+        public PrincipalForm()
         {
             InitializeComponent();
         }
