+++ Control Bibliotecario/Control Bibliotecario/UI/RegistroForm.resx
